requirements.txt
fastapi
uvicorn
pandas
numpy
openpyxl
python-dotenv
google-generativeai